refactor(backend): move function to dependent utils
