+++ crates/services/miscellaneous/src/lib.rs
@@ -52,7 +52,7 @@ use dependent_utils::{
     first_metadata_image_as_url, get_entity_recently_consumed, get_metadata_provider,
     get_openlibrary_service, get_tmdb_non_media_service, get_users_and_cte_monitoring_entity,
     get_users_monitoring_entity, handle_after_media_seen_tasks, is_metadata_finished_by_user,
-    metadata_images_as_urls, post_review, progress_update,
+    metadata_images_as_urls, metadata_list, post_review, progress_update,
     refresh_collection_to_entity_association, remove_entity_from_collection,
     update_metadata_and_notify_users,
 };
@@ -68,9 +68,9 @@ use media_models::{
     CreateReviewCommentInput, GenreDetailsInput, GenreListItem, GraphqlCalendarEvent,
     GraphqlMediaAssets, GraphqlMetadataDetails, GraphqlMetadataGroup, GraphqlVideoAsset,
     GroupedCalendarEvent, ImportOrExportItemReviewComment, MarkEntityAsPartialInput,
-    MediaAssociatedPersonStateChanges, MediaGeneralFilter, MediaSortBy, MetadataCreator,
-    MetadataCreatorGroupedByRole, MetadataFreeCreator, MetadataGroupsListInput, MetadataImage,
-    MetadataListInput, MetadataPartialDetails, MetadataVideo, MetadataVideoSource, PartialMetadata,
+    MediaAssociatedPersonStateChanges, MetadataCreator, MetadataCreatorGroupedByRole,
+    MetadataFreeCreator, MetadataGroupsListInput, MetadataImage, MetadataListInput,
+    MetadataPartialDetails, MetadataVideo, MetadataVideoSource, PartialMetadata,
     PartialMetadataWithoutId, PeopleListInput, PersonAndMetadataGroupsSortBy,
     PersonDetailsGroupedByRole, PersonDetailsItemWithCharacter, PodcastSpecifics,
     ProgressUpdateInput, ReviewPostedEvent, SeenAnimeExtraInformation, SeenPodcastExtraInformation,
@@ -80,8 +80,7 @@ use media_models::{
     UserUpcomingCalendarEventInput,
 };
 use migrations::{
-    AliasedCalendarEvent, AliasedMetadata, AliasedMetadataToGenre, AliasedReview, AliasedSeen,
-    AliasedUserToEntity,
+    AliasedCalendarEvent, AliasedMetadata, AliasedMetadataToGenre, AliasedSeen, AliasedUserToEntity,
 };
 use nanoid::nanoid;
 use notification_service::send_notification;
@@ -94,19 +93,19 @@ use providers::{
 use rust_decimal::Decimal;
 use rust_decimal_macros::dec;
 use sea_orm::{
-    prelude::DateTimeUtc, query::UpdateMany, sea_query::NullOrdering, ActiveModelTrait,
-    ActiveValue, ColumnTrait, ConnectionTrait, DatabaseBackend, DatabaseConnection, EntityTrait,
-    FromQueryResult, ItemsAndPagesNumber, JoinType, ModelTrait, Order, PaginatorTrait, QueryFilter,
-    QueryOrder, QuerySelect, QueryTrait, RelationTrait, Statement, TransactionTrait,
+    prelude::DateTimeUtc, query::UpdateMany, ActiveModelTrait, ActiveValue, ColumnTrait,
+    ConnectionTrait, DatabaseBackend, DatabaseConnection, EntityTrait, FromQueryResult,
+    ItemsAndPagesNumber, JoinType, ModelTrait, Order, PaginatorTrait, QueryFilter, QueryOrder,
+    QuerySelect, QueryTrait, RelationTrait, Statement, TransactionTrait,
 };
 use sea_query::{
-    extension::postgres::PgExpr, Alias, Asterisk, Cond, Condition, Expr, Func, PgFunc,
+    extension::postgres::PgExpr, Alias, Asterisk, Condition, Expr, Func, PgFunc,
     PostgresQueryBuilder, Query, SelectStatement,
 };
 use serde::{Deserialize, Serialize};
 use supporting_service::SupportingService;
 use traits::{MediaProvider, TraceOk};
-use user_models::{DashboardElementLot, UserReviewScale};
+use user_models::DashboardElementLot;
 use uuid::Uuid;
 
 static ENTITY_UPDATE_CHUNK_SIZE: usize = 5;
@@ -893,151 +892,7 @@ ORDER BY RANDOM() LIMIT 10;
         user_id: String,
         input: MetadataListInput,
     ) -> Result<SearchResults<String>> {
-        let preferences = user_by_id(&user_id, &self.0).await?.preferences;
-
-        let avg_rating_col = "user_average_rating";
-        let cloned_user_id_1 = user_id.clone();
-        let cloned_user_id_2 = user_id.clone();
-
-        let order_by = input
-            .sort
-            .clone()
-            .map(|a| graphql_to_db_order(a.order))
-            .unwrap_or(Order::Asc);
-        let review_scale = match preferences.general.review_scale {
-            UserReviewScale::OutOfFive => 20,
-            UserReviewScale::OutOfHundred | UserReviewScale::ThreePointSmiley => 1,
-        };
-        let take = input.take.unwrap_or(PAGE_SIZE as u64);
-        let page: u64 = input
-            .search
-            .clone()
-            .and_then(|s| s.page)
-            .unwrap_or(1)
-            .try_into()
-            .unwrap();
-        let paginator = Metadata::find()
-            .select_only()
-            .column(metadata::Column::Id)
-            .expr_as(
-                Func::round_with_precision(
-                    Func::avg(
-                        Expr::col((AliasedReview::Table, AliasedReview::Rating)).div(review_scale),
-                    ),
-                    review_scale,
-                ),
-                avg_rating_col,
-            )
-            .group_by(metadata::Column::Id)
-            .group_by(user_to_entity::Column::MediaReason)
-            .filter(user_to_entity::Column::UserId.eq(&user_id))
-            .apply_if(input.lot, |query, v| {
-                query.filter(metadata::Column::Lot.eq(v))
-            })
-            .inner_join(UserToEntity)
-            .join(
-                JoinType::LeftJoin,
-                metadata::Relation::Review
-                    .def()
-                    .on_condition(move |_left, right| {
-                        Condition::all().add(
-                            Expr::col((right, review::Column::UserId)).eq(cloned_user_id_1.clone()),
-                        )
-                    }),
-            )
-            .join(
-                JoinType::LeftJoin,
-                metadata::Relation::Seen
-                    .def()
-                    .on_condition(move |_left, right| {
-                        Condition::all().add(
-                            Expr::col((right, seen::Column::UserId)).eq(cloned_user_id_2.clone()),
-                        )
-                    }),
-            )
-            .apply_if(input.search.and_then(|s| s.query), |query, v| {
-                query.filter(
-                    Cond::any()
-                        .add(Expr::col(metadata::Column::Title).ilike(ilike_sql(&v)))
-                        .add(Expr::col(metadata::Column::Description).ilike(ilike_sql(&v))),
-                )
-            })
-            .apply_if(
-                input.filter.clone().and_then(|f| f.collections),
-                |query, v| {
-                    apply_collection_filter(
-                        query,
-                        Some(v),
-                        input.invert_collection,
-                        metadata::Column::Id,
-                        collection_to_entity::Column::MetadataId,
-                    )
-                },
-            )
-            .apply_if(input.filter.and_then(|f| f.general), |query, v| match v {
-                MediaGeneralFilter::All => query.filter(metadata::Column::Id.is_not_null()),
-                MediaGeneralFilter::Rated => query.filter(review::Column::Id.is_not_null()),
-                MediaGeneralFilter::Unrated => query.filter(review::Column::Id.is_null()),
-                MediaGeneralFilter::Unfinished => query.filter(
-                    Expr::expr(
-                        Expr::val(UserToMediaReason::Finished.to_string())
-                            .eq(PgFunc::any(Expr::col(user_to_entity::Column::MediaReason))),
-                    )
-                    .not(),
-                ),
-                s => query.filter(seen::Column::State.eq(match s {
-                    MediaGeneralFilter::Dropped => SeenState::Dropped,
-                    MediaGeneralFilter::OnAHold => SeenState::OnAHold,
-                    _ => unreachable!(),
-                })),
-            })
-            .apply_if(input.sort.map(|s| s.by), |query, v| match v {
-                MediaSortBy::LastUpdated => query
-                    .order_by(user_to_entity::Column::LastUpdatedOn, order_by)
-                    .group_by(user_to_entity::Column::LastUpdatedOn),
-                MediaSortBy::Title => query.order_by(metadata::Column::Title, order_by),
-                MediaSortBy::ReleaseDate => query.order_by_with_nulls(
-                    metadata::Column::PublishYear,
-                    order_by,
-                    NullOrdering::Last,
-                ),
-                MediaSortBy::LastSeen => query.order_by_with_nulls(
-                    seen::Column::FinishedOn.max(),
-                    order_by,
-                    NullOrdering::Last,
-                ),
-                MediaSortBy::UserRating => query.order_by_with_nulls(
-                    Expr::col(Alias::new(avg_rating_col)),
-                    order_by,
-                    NullOrdering::Last,
-                ),
-                MediaSortBy::ProviderRating => query.order_by_with_nulls(
-                    metadata::Column::ProviderRating,
-                    order_by,
-                    NullOrdering::Last,
-                ),
-            })
-            .into_tuple::<String>()
-            .paginate(&self.0.db, take);
-        let ItemsAndPagesNumber {
-            number_of_items,
-            number_of_pages,
-        } = paginator.num_items_and_pages().await?;
-        let mut items = vec![];
-        for c in paginator.fetch_page(page - 1).await? {
-            items.push(c);
-        }
-        Ok(SearchResults {
-            details: SearchDetails {
-                total: number_of_items.try_into().unwrap(),
-                next_page: if page < number_of_pages {
-                    Some((page + 1).try_into().unwrap())
-                } else {
-                    None
-                },
-            },
-            items,
-        })
+        metadata_list(user_id, input, &self.0).await
     }
 
     pub async fn deploy_bulk_progress_update(
